Move config file detection code to config module
